requirements.txt
google-genai>=1.0.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
Pillow>=10.0.0
click>=8.0.0
